Refactor: split widget export modal content
Fixes #21881
